@@ -1,5 +1,10 @@
 module github.com/TencentBlueKing/ci-repoAnalysis/analysis-tool-sdk-golang
 
-go 1.20
+go 1.21.0
 
-require golang.org/x/sync v0.3.0 // indirect
+require (
+	github.com/hashicorp/go-retryablehttp v0.7.4
+	golang.org/x/sync v0.3.0
+)
+
+require github.com/hashicorp/go-cleanhttp v0.5.2 // indirect
